fix(docker): update Docker modules for bollard 0.16 API compatibility
- RestartContainerOptions.t: i64 -> isize
- BuildImageOptions: remove non-existent target field
- import_image: replace hyper::Body with bytes::Bytes
- CreateNetworkOptions/CreateVolumeOptions: use String type parameter
- CreateNetworkOptions.ipam: unwrap Option<Ipam> to Ipam

+++ src/modules/docker/docker_container.rs
@@ -564,7 +564,7 @@ impl DockerContainerModule {
         timeout: Option<i64>,
     ) -> ModuleResult<()> {
         let options = RestartContainerOptions {
-            t: timeout.unwrap_or(10) as i64,
+            t: timeout.unwrap_or(10) as isize,
         };
         docker
             .restart_container(name, Some(options))
